chore: convert commons.aria to ES Modules
