+++ .eslintrc
@@ -74,6 +74,7 @@
   "overrides": [
     {
       "files": [
+        "lib/commons/aria/*.js",
         "lib/commons/forms/*.js",
         "lib/commons/matches/*.js",
         "lib/commons/utils/*.js"
