+++ lib/commons/aria/arialabel-text.js
@@ -1,5 +1,3 @@
-/* global aria */
-
 /**
  * Get the text value of aria-label, if any
  *
@@ -7,12 +5,16 @@
  * @param {VirtualNode|Element} element
  * @return {string} ARIA label
  */
-aria.arialabelText = function arialabelText(node) {
+function arialabelText(node) {
+	// TODO: es-module-AbstractVirtualNode
 	if (node instanceof axe.AbstractVirtualNode === false) {
 		if (node.nodeType !== 1) {
 			return '';
 		}
+		// TODO: es-module-utils.getNodeFromTree
 		node = axe.utils.getNodeFromTree(node);
 	}
 	return node.attr('aria-label') || '';
-};
+}
+
+export default arialabelText;
