+++ lib/commons/aria/arialabelledby-text.js
@@ -1,4 +1,4 @@
-/* global aria, dom, text */
+/* global dom, text */
 
 /**
  * Get the accessible name based on aria-labelledby
@@ -11,7 +11,7 @@
  * @property {Bool} debug Enable logging for formControlValue
  * @return {string} Cancatinated text value for referenced elements
  */
-aria.arialabelledbyText = function arialabelledbyText(node, context = {}) {
+function arialabelledbyText(node, context = {}) {
 	node = node.actualNode || node;
 	/**
 	 * Note: The there are significant difference in how many "leads" browsers follow.
@@ -34,8 +34,10 @@ aria.arialabelledbyText = function arialabelledbyText(node, context = {}) {
 		return '';
 	}
 
+	// TODO: es-module-dom.idrefs
 	const refs = dom.idrefs(node, 'aria-labelledby').filter(elm => elm);
 	return refs.reduce((accessibleName, elm) => {
+		// TODO: es-module-text.accessibleText
 		const accessibleNameAdd = text.accessibleText(elm, {
 			// Prevent the infinite reference loop:
 			inLabelledByContext: true,
@@ -49,4 +51,6 @@ aria.arialabelledbyText = function arialabelledbyText(node, context = {}) {
 			return `${accessibleName} ${accessibleNameAdd}`;
 		}
 	}, '');
-};
+}
+
+export default arialabelledbyText;
