+++ lib/commons/aria/get-element-unallowed-roles.js
@@ -1,4 +1,7 @@
-/* global aria */
+import isValidRole from './is-valid-role';
+import getImplicitRole from './implicit-role';
+import getRoleType from './get-role-type';
+import isAriaRoleAllowedOnElement from './is-aria-role-allowed-on-element';
 
 // dpub roles which are subclassing roles that are implicit on some native
 // HTML elements (img, link, etc.)
@@ -29,6 +32,7 @@ function getRoleSegments(node) {
 	}
 
 	if (node.hasAttribute('role')) {
+		// TODO: es-module-utils.tokenList
 		const nodeRoles = axe.utils.tokenList(
 			node.getAttribute('role').toLowerCase()
 		);
@@ -36,6 +40,7 @@ function getRoleSegments(node) {
 	}
 
 	if (node.hasAttributeNS('http://www.idpf.org/2007/ops', 'type')) {
+		// TODO: es-module-utils.tokenList
 		const epubRoles = axe.utils
 			.tokenList(
 				node
@@ -48,7 +53,7 @@ function getRoleSegments(node) {
 	}
 
 	// filter invalid roles
-	roles = roles.filter(role => axe.commons.aria.isValidRole(role));
+	roles = roles.filter(role => isValidRole(role));
 
 	return roles;
 }
@@ -61,19 +66,17 @@ function getRoleSegments(node) {
  * @param {String} allowImplicit option to allow implicit roles, defaults to true
  * @return {Array<String>} retruns an array of roles that are not allowed on the given node
  */
-aria.getElementUnallowedRoles = function getElementUnallowedRoles(
-	node,
-	allowImplicit = true
-) {
+function getElementUnallowedRoles(node, allowImplicit = true) {
 	const tagName = node.nodeName.toUpperCase();
 
 	// by pass custom elements
+	// TODO: es-module-utils.isHtmlElement
 	if (!axe.utils.isHtmlElement(node)) {
 		return [];
 	}
 
 	const roleSegments = getRoleSegments(node);
-	const implicitRole = axe.commons.aria.implicitRole(node);
+	const implicitRole = getImplicitRole(node);
 
 	// stores all roles that are not allowed for a specific element most often an element only has one explicit role
 	const unallowedRoles = roleSegments.filter(role => {
@@ -86,7 +89,7 @@ aria.getElementUnallowedRoles = function getElementUnallowedRoles(
 		// if role is a dpub role make sure it's used on an element with a valid
 		// implicit role fallback
 		if (allowImplicit && dpubRoles.includes(role)) {
-			const roleType = axe.commons.aria.getRoleType(role);
+			const roleType = getRoleType(role);
 			if (implicitRole !== roleType) {
 				return true;
 			}
@@ -99,6 +102,7 @@ aria.getElementUnallowedRoles = function getElementUnallowedRoles(
 			!(
 				role === 'row' &&
 				tagName === 'TR' &&
+				// TODO: es-module-utils.matchesSelector
 				axe.utils.matchesSelector(node, 'table[role="grid"] > tr')
 			)
 		) {
@@ -106,8 +110,10 @@ aria.getElementUnallowedRoles = function getElementUnallowedRoles(
 		}
 
 		// check if role is allowed on element
-		return !aria.isAriaRoleAllowedOnElement(node, role);
+		return !isAriaRoleAllowedOnElement(node, role);
 	});
 
 	return unallowedRoles;
-};
+}
+
+export default getElementUnallowedRoles;
