+++ lib/commons/aria/get-owned-virtual.js
@@ -1,4 +1,4 @@
-/* global aria, dom */
+/* global dom */
 
 /**
  * Get an element's owned elements
@@ -6,7 +6,7 @@
  * @param {VirtualNode} element
  * @return {VirtualNode[]} Owned elements
  */
-aria.getOwnedVirtual = function getOwned({ actualNode, children }) {
+function getOwnedVirtual({ actualNode, children }) {
 	if (!actualNode || !children) {
 		throw new Error('getOwnedVirtual requires a virtual node');
 	}
@@ -14,11 +14,15 @@ aria.getOwnedVirtual = function getOwned({ actualNode, children }) {
 	// TODO: Descend into children with role=presentation|none
 	// TODO: Exclude descendents owned by other elements
 
+	// TODO: es-module-dom.idrefs
 	return dom.idrefs(actualNode, 'aria-owns').reduce((ownedElms, element) => {
 		if (element) {
+			// TODO: es-module-utils.getNodeFromTree
 			const virtualNode = axe.utils.getNodeFromTree(element);
 			ownedElms.push(virtualNode);
 		}
 		return ownedElms;
 	}, children);
-};
+}
+
+export default getOwnedVirtual;
