+++ lib/commons/aria/get-role.js
@@ -1,4 +1,5 @@
-/* global aria, axe */
+import isValidRole from './is-valid-role';
+import getImplicitRole from './implicit-role';
 
 /**
  * Return the accessible role of an element
@@ -14,15 +15,13 @@
  * @param {boolean} options.dpub  Allow role to be any (valid) doc-* roles
  * @returns {string|null} Role or null
  */
-aria.getRole = function getRole(
-	node,
-	{ noImplicit, fallback, abstracts, dpub } = {}
-) {
+function getRole(node, { noImplicit, fallback, abstracts, dpub } = {}) {
 	node = node.actualNode || node;
 	if (node.nodeType !== 1) {
 		return null;
 	}
 	const roleAttr = (node.getAttribute('role') || '').trim().toLowerCase();
+	// TODO: es-module-utils.tokenList
 	const roleList = fallback ? axe.utils.tokenList(roleAttr) : [roleAttr];
 
 	// Get the first valid role:
@@ -30,15 +29,17 @@ aria.getRole = function getRole(
 		if (!dpub && role.substr(0, 4) === 'doc-') {
 			return false;
 		}
-		return aria.isValidRole(role, { allowAbstract: abstracts });
+		return isValidRole(role, { allowAbstract: abstracts });
 	});
 
 	const explicitRole = validRoles[0];
 
 	// Get the implicit role, if permitted
 	if (!explicitRole && !noImplicit) {
-		return aria.implicitRole(node);
+		return getImplicitRole(node);
 	}
 
 	return explicitRole || null;
-};
+}
+
+export default getRole;
