+++ lib/commons/aria/is-accessible-ref.js
@@ -1,8 +1,11 @@
-/* global aria, axe, dom */
+/* global dom */
+import lookupTable from './lookup-table';
+
 const idRefsRegex = /^idrefs?$/;
 
 function cacheIdRefs(node, refAttrs) {
 	if (node.hasAttribute) {
+		// TODO: es-module-_cache
 		const idRefs = axe._cache.get('idRefs');
 
 		if (node.nodeName.toUpperCase() === 'LABEL' && node.hasAttribute('for')) {
@@ -18,6 +21,7 @@ function cacheIdRefs(node, refAttrs) {
 
 			const attrValue = node.getAttribute(attr);
 
+			// TODO: es-module-utils.tokenList
 			const tokens = axe.utils.tokenList(attrValue);
 
 			for (let k = 0; k < tokens.length; ++k) {
@@ -36,24 +40,29 @@ function cacheIdRefs(node, refAttrs) {
  * @param {Element} node
  * @returns {Boolean}
  */
-aria.isAccessibleRef = function isAccessibleRef(node) {
+function isAccessibleRef(node) {
 	node = node.actualNode || node;
+	// TODO: es-module-dom.getRootNode
 	let root = dom.getRootNode(node);
 	root = root.documentElement || root; // account for shadow roots
 	const id = node.id;
 
 	// because axe.commons is not available in axe.utils, we can't do
 	// this caching when we build up the virtual tree
+	// TODO: es-module-_cache
 	if (!axe._cache.get('idRefs')) {
 		axe._cache.set('idRefs', {});
 		// Get all idref(s) attributes on the lookup table
-		const refAttrs = Object.keys(aria.lookupTable.attributes).filter(attr => {
-			const { type } = aria.lookupTable.attributes[attr];
+		const refAttrs = Object.keys(lookupTable.attributes).filter(attr => {
+			const { type } = lookupTable.attributes[attr];
 			return idRefsRegex.test(type);
 		});
 
 		cacheIdRefs(root, refAttrs);
 	}
 
+	// TODO: es-module-_cache
 	return axe._cache.get('idRefs')[id] === true;
-};
+}
+
+export default isAccessibleRef;
