+++ lib/commons/aria/is-aria-role-allowed-on-element.js
@@ -1,4 +1,6 @@
-/* global aria, matches */
+import lookupTable from './lookup-table';
+import matches from '../matches';
+
 /**
  * @description validate if a given role is an allowed ARIA role for the supplied node
  * @method isAriaRoleAllowedOnElement
@@ -6,12 +8,8 @@
  * @param {String} role aria role to check
  * @return {Boolean} retruns true/false
  */
-aria.isAriaRoleAllowedOnElement = function isAriaRoleAllowedOnElement(
-	node,
-	role
-) {
+function isAriaRoleAllowedOnElement(node, role) {
 	const nodeName = node.nodeName.toUpperCase();
-	const lookupTable = axe.commons.aria.lookupTable;
 
 	// if given node can have no role - return false
 	if (matches(node, lookupTable.elementsAllowedNoRole)) {
@@ -38,4 +36,6 @@ aria.isAriaRoleAllowedOnElement = function isAriaRoleAllowedOnElement(
 		return lookupTable.evaluateRoleForElement[nodeName]({ node, role, out });
 	}
 	return out;
-};
+}
+
+export default isAriaRoleAllowedOnElement;
