+++ lib/commons/aria/is-unsupported-role.js
@@ -1,4 +1,4 @@
-/* global aria */
+import lookupTable from './lookup-table';
 
 /**
  * Check if a given role is unsupported
@@ -8,7 +8,9 @@
  * @param {String} role The role to check
  * @return {Boolean}
  */
-aria.isUnsupportedRole = function(role) {
-	const roleDefinition = aria.lookupTable.role[role];
+function isUnsupportedRole(role) {
+	const roleDefinition = lookupTable.role[role];
 	return roleDefinition ? roleDefinition.unsupported : false;
-};
+}
+
+export default isUnsupportedRole;
