+++ lib/commons/aria/label-virtual.js
@@ -1,4 +1,4 @@
-/* global axe, aria, dom, text */
+/* global dom, text */
 
 /**
  * Gets the accessible ARIA label text of a given element
@@ -9,15 +9,18 @@
  * @param  {Object} actualNode The virtualNode to test
  * @return {Mixed}  String of visible text, or `null` if no label is found
  */
-aria.labelVirtual = function({ actualNode }) {
+function labelVirtual({ actualNode }) {
 	let ref, candidate;
 
 	if (actualNode.getAttribute('aria-labelledby')) {
 		// aria-labelledby
+		// TODO: es-module-dom.idrefs
 		ref = dom.idrefs(actualNode, 'aria-labelledby');
 		candidate = ref
 			.map(function(thing) {
+				// TODO: es-module-utils.getNodeFromTree
 				const vNode = axe.utils.getNodeFromTree(thing);
+				// TODO: es-module-text.visibleVirtual
 				return vNode ? text.visibleVirtual(vNode, true) : '';
 			})
 			.join(' ')
@@ -31,6 +34,7 @@ aria.labelVirtual = function({ actualNode }) {
 	// aria-label
 	candidate = actualNode.getAttribute('aria-label');
 	if (candidate) {
+		// TODO: es-module-text.sanitize
 		candidate = text.sanitize(candidate).trim();
 		if (candidate) {
 			return candidate;
@@ -38,17 +42,6 @@ aria.labelVirtual = function({ actualNode }) {
 	}
 
 	return null;
-};
+}
 
-/**
- * Gets the aria label for a given node
- * @method label
- * @memberof axe.commons.aria
- * @instance
- * @param  {HTMLElement} node The element to check
- * @return {Mixed} String of visible text, or `null` if no label is found
- */
-aria.label = function(node) {
-	node = axe.utils.getNodeFromTree(node);
-	return aria.labelVirtual(node);
-};
+export default labelVirtual;
