+++ lib/commons/aria/named-from-contents.js
@@ -1,4 +1,5 @@
-/* global aria */
+import getRole from './get-role';
+import lookupTable from './lookup-table';
 
 /**
  * Check if an element is named from contents
@@ -8,14 +9,14 @@
  * @property {Bool} strict Whether or not to follow the spects strictly
  * @return {Bool}
  */
-aria.namedFromContents = function namedFromContents(node, { strict } = {}) {
+function namedFromContents(node, { strict } = {}) {
 	node = node.actualNode || node;
 	if (node.nodeType !== 1) {
 		return false;
 	}
 
-	const role = aria.getRole(node);
-	const roleDef = aria.lookupTable.role[role];
+	const role = getRole(node);
+	const roleDef = lookupTable.role[role];
 
 	if (
 		(roleDef && roleDef.nameFrom.includes('contents')) ||
@@ -36,4 +37,6 @@ aria.namedFromContents = function namedFromContents(node, { strict } = {}) {
 		return false;
 	}
 	return !roleDef || ['presentation', 'none'].includes(role);
-};
+}
+
+export default namedFromContents;
