+++ lib/commons/aria/validate-attr-value.js
@@ -1,4 +1,5 @@
-/* global aria, dom */
+/* global dom */
+import lookupTable from './lookup-table';
 
 /**
  * Validate the value of an ARIA attribute
@@ -9,13 +10,14 @@
  * @param  {String} attr The name of the attribute
  * @return {Boolean}
  */
-aria.validateAttrValue = function validateAttrValue(node, attr) {
+function validateAttrValue(node, attr) {
 	'use strict';
-	var matches,
-		list,
-		value = node.getAttribute(attr),
-		attrInfo = aria.lookupTable.attributes[attr];
+	let matches;
+	let list;
+	const value = node.getAttribute(attr);
+	const attrInfo = lookupTable.attributes[attr];
 
+	// TODO: es-module-dom.getRootNode
 	var doc = dom.getRootNode(node);
 
 	if (!attrInfo) {
@@ -34,6 +36,7 @@ aria.validateAttrValue = function validateAttrValue(node, attr) {
 			);
 
 		case 'nmtokens':
+			// TODO: es-module-utils.tokenList
 			list = axe.utils.tokenList(value);
 			// Check if any value isn't in the list of values
 			return list.reduce(function(result, token) {
@@ -45,6 +48,7 @@ aria.validateAttrValue = function validateAttrValue(node, attr) {
 			return !!(value && doc.getElementById(value));
 
 		case 'idrefs':
+			// TODO: es-module-utils.tokenList
 			list = axe.utils.tokenList(value);
 			return list.some(token => doc.getElementById(token));
 
@@ -59,4 +63,6 @@ aria.validateAttrValue = function validateAttrValue(node, attr) {
 		case 'int':
 			return /^[-+]?[0-9]+$/.test(value);
 	}
-};
+}
+
+export default validateAttrValue;
