+++ lib/commons/index.js
@@ -9,4 +9,5 @@
  */
 
 var commons = {};
+var aria = {};
 var matches;
