+++ lib/commons/matches/index.js
@@ -26,6 +26,7 @@ matches.fromPrimative = fromPrimative;
 matches.nodeName = nodeName;
 matches.properties = properties;
 
-// TODO: es-module-commons. convert to:
-// export default matches
+// TODO: es-module-commons. remove
 commons.matches = matchesFn;
+
+export default matches;
